[C8] Split loft_skin.go by responsibility (#2236)
model/feature/loft_skin.go was 1108 LOC, 2.2x the 500-line ceiling. Split the
loft skinning builder by responsibility; each file is now under the ceiling:

- loft_skin.go (185): section preparation/alignment — winding match, best
  rotational offset, closure shift, spline resampling of the section loops.
- loft_skin_rails.go (227): rail and guide handling — rail/map/area-graph/
  centreline guides re-tracking the sections with a falloff.
- loft_skin_continuity.go (306): the tangency/continuity conditions — the
  per-section tangent/G2/G3 derivation and the apex/angle/face end conditions.
- loft_skin_fit.go (373): the skin surface fit — the hermite blend (C0/C1/G2/
  G3), adaptive segment/around sampling, and the hermite3/5/7 basis.
- loft_skin_cap.go (67): cap construction — extending the end loops outward.

Behaviour-preserving: split programmatically by top-level declaration,
verified every non-comment code line is byte-identical to the original (only
per-file import redistribution differs); imports resolved per file with
goimports.

The .golangci.yml funlen/gocyclo exclude-rule for model/feature/loft_skin.go
is broadened to loft_skin[_a-z0-9]*.go (same as the dressup broadening) so the
split files inherit the SAME pre-existing exclusion. Feature package builds,
archguard passes, lint 0. Net delta of recognizers / tolerance constants /
fallback sites / type assertions: 0.

Part of #2183 (C8: structure/size limits). Closes #2236.

(cherry picked from commit edc4008176ae263cffe38edc4e52fbda11613b9b)
